test: route slow cmd/gc cases out of fast loop
