+++ cmd/gc/beads_provider_lifecycle_test.go
@@ -2120,6 +2120,7 @@ func TestStartBeadsLifecycleDoesNotMutateProcessDoltEnv(t *testing.T) {
 }
 
 func TestGcBeadsBdStartUsesRootBeadsDataDir(t *testing.T) {
+	skipSlowCmdGCTest(t, "starts the real gc-beads-bd lifecycle script; run make test-cmd-gc-process for full coverage")
 	doltPath, err := exec.LookPath("dolt")
 	if err != nil {
 		t.Skip("dolt not installed")
@@ -2626,6 +2627,7 @@ exit 2
 }
 
 func TestHealthBeadsProviderPublishesManagedRuntimeStateWhenHealthyButUnpublished(t *testing.T) {
+	skipSlowCmdGCTest(t, "starts the real gc-beads-bd lifecycle script; run make test-cmd-gc-process for full coverage")
 	cityPath, _ := setupManagedBdWaitTestCity(t)
 
 	if err := os.Remove(managedDoltStatePath(cityPath)); err != nil && !os.IsNotExist(err) {
@@ -3777,98 +3779,6 @@ esac
 	}
 }
 
-func listenOnRandomPort(t *testing.T) net.Listener {
-	t.Helper()
-	ln, err := net.Listen("tcp", "127.0.0.1:0")
-	if err != nil {
-		t.Fatalf("listen: %v", err)
-	}
-	return ln
-}
-
-func reserveRandomTCPPort(t *testing.T) int {
-	t.Helper()
-	ln := listenOnRandomPort(t)
-	port := ln.Addr().(*net.TCPAddr).Port
-	_ = ln.Close()
-	return port
-}
-
-func startTCPListenerProcess(t *testing.T, port int) *exec.Cmd {
-	t.Helper()
-	cmd := exec.Command("python3", "-c", `
-import signal
-import socket
-import sys
-import time
-port = int(sys.argv[1])
-sock = socket.socket()
-sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-sock.bind(("127.0.0.1", port))
-sock.listen(5)
-def _stop(*_args):
-    raise SystemExit(0)
-signal.signal(signal.SIGTERM, _stop)
-signal.signal(signal.SIGINT, _stop)
-while True:
-    time.sleep(1)
-`, strconv.Itoa(port))
-	if err := cmd.Start(); err != nil {
-		t.Fatalf("start listener process: %v", err)
-	}
-	t.Cleanup(func() {
-		if cmd.Process != nil {
-			_ = cmd.Process.Kill()
-		}
-		_ = cmd.Wait()
-	})
-	deadline := time.Now().Add(5 * time.Second)
-	for time.Now().Before(deadline) {
-		conn, err := net.DialTimeout("tcp", fmt.Sprintf("127.0.0.1:%d", port), 200*time.Millisecond)
-		if err == nil {
-			_ = conn.Close()
-			return cmd
-		}
-		time.Sleep(50 * time.Millisecond)
-	}
-	t.Fatalf("listener process on %d did not become ready", port)
-	return nil
-}
-
-func writeDoltState(cityPath string, state doltRuntimeState) error {
-	stateDir := filepath.Join(cityPath, ".gc", "runtime", "packs", "dolt")
-	if err := os.MkdirAll(stateDir, 0o755); err != nil {
-		return err
-	}
-	data := fmt.Sprintf(`{"running":%t,"pid":%d,"port":%d,"data_dir":%q,"started_at":%q}`,
-		state.Running, state.PID, state.Port, state.DataDir, state.StartedAt)
-	return os.WriteFile(filepath.Join(stateDir, "dolt-state.json"), []byte(data), 0o644)
-}
-
-// writeTestScript creates a shell script that exits with the given code.
-// If stderrMsg is non-empty, the script writes it to stderr before exiting.
-func writeTestScript(t *testing.T, _ string, exitCode int, stderrMsg string) string {
-	t.Helper()
-	dir := t.TempDir()
-	script := filepath.Join(dir, "test-beads.sh")
-
-	content := "#!/bin/sh\n"
-	if stderrMsg != "" {
-		content += "echo '" + stderrMsg + "' >&2\n"
-	}
-	content += "exit " + itoa(exitCode) + "\n"
-
-	if err := os.WriteFile(script, []byte(content), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	return script
-}
-
-// itoa is a simple int to string converter for test scripts.
-func itoa(n int) string {
-	return []string{"0", "1", "2"}[n]
-}
-
 // ── isExternalDolt tests ──────────────────────────────────────────────
 
 func TestIsExternalDoltEnvFallback(t *testing.T) {
@@ -4122,6 +4032,7 @@ dolt.port: 3307
 }
 
 func TestGcBeadsBdStartIgnoresReachableCompatPortFileInput(t *testing.T) {
+	skipSlowCmdGCTest(t, "starts the real gc-beads-bd lifecycle script; run make test-cmd-gc-process for full coverage")
 	cityPath := t.TempDir()
 	if err := os.MkdirAll(filepath.Join(cityPath, ".gc"), 0o755); err != nil {
 		t.Fatal(err)
@@ -4671,6 +4582,7 @@ esac
 }
 
 func TestGcBeadsBdStartDoesNotReplaceLiveLockFileInode(t *testing.T) {
+	skipSlowCmdGCTest(t, "starts the real gc-beads-bd lifecycle script; run make test-cmd-gc-process for full coverage")
 	if _, err := exec.LookPath("flock"); err != nil {
 		t.Skip("flock not installed")
 	}
@@ -4846,6 +4758,7 @@ while [ ! -f "$release_file" ]; do
 }
 
 func TestGcBeadsBdStartWaitsForConcurrentStarterSuccess(t *testing.T) {
+	skipSlowCmdGCTest(t, "starts the real gc-beads-bd lifecycle script; run make test-cmd-gc-process for full coverage")
 	cityPath := t.TempDir()
 	if err := MaterializeBuiltinPacks(cityPath); err != nil {
 		t.Fatalf("MaterializeBuiltinPacks: %v", err)
@@ -5388,6 +5301,7 @@ func readManagedDoltConfigForTest(t *testing.T, path string) managedDoltConfigFo
 }
 
 func TestGcBeadsBdStartIsIdempotentWhenAlreadyRunning(t *testing.T) {
+	skipSlowCmdGCTest(t, "starts the real gc-beads-bd lifecycle script; run make test-cmd-gc-process for full coverage")
 	cityPath := t.TempDir()
 	if err := os.MkdirAll(filepath.Join(cityPath, ".gc"), 0o755); err != nil {
 		t.Fatal(err)
@@ -5516,6 +5430,7 @@ esac
 }
 
 func TestGcBeadsBdStartRestartsServerHoldingDeletedDataInodes(t *testing.T) {
+	skipSlowCmdGCTest(t, "starts the real gc-beads-bd lifecycle script; run make test-cmd-gc-process for full coverage")
 	cityPath := t.TempDir()
 	if err := os.MkdirAll(filepath.Join(cityPath, ".gc"), 0o755); err != nil {
 		t.Fatal(err)
@@ -5648,6 +5563,7 @@ esac
 }
 
 func TestGcBeadsBdEnsureReadyDoesNotRestartAfterTransientTCPProbeFailure(t *testing.T) {
+	skipSlowCmdGCTest(t, "starts the real gc-beads-bd lifecycle script; run make test-cmd-gc-process for full coverage")
 	cityPath := t.TempDir()
 	if err := os.MkdirAll(filepath.Join(cityPath, ".gc"), 0o755); err != nil {
 		t.Fatal(err)
@@ -6409,6 +6325,7 @@ func TestNormalizeCanonicalBdScopeFilesRepairsCityAndRigScopeFiles(t *testing.T)
 }
 
 func TestGcBeadsBdInitNormalizesScopeAndRemovesLocalServerArtifacts(t *testing.T) {
+	skipSlowCmdGCTest(t, "starts the real gc-beads-bd lifecycle script; run make test-cmd-gc-process for full coverage")
 	cityPath := t.TempDir()
 	rigPath := filepath.Join(cityPath, "frontend")
 	if err := os.MkdirAll(filepath.Join(cityPath, ".gc"), 0o755); err != nil {
