+++ cmd/gc/live_submit_probe_test.go
@@ -20,6 +20,7 @@ import (
 
 func preferRealBDOnPath(t *testing.T) {
 	t.Helper()
+	skipSlowCmdGCTest(t, "requires a live bd-managed session probe; run without -short")
 
 	currentPath := os.Getenv("PATH")
 	pathEntries := filepath.SplitList(currentPath)
